requirements.txt
pandas
numpy
yfinance
pillow
tqdm
tensorflow
matplotlib
imblearn
scikit-learn
bayes-opt